Edit classes for the registration process
