--- server/src/main/java/io/javaclasses/filehub/server/RegistrationCommand.java
@@ -1,12 +1,0 @@
-package io.javaclasses.filehub.server;
-
-public class RegistrationCommand {
-
-    public void register(UserCredentials userCredentials) throws ValidationError {
-
-        if (userCredentials.getUsername().getValue().equals("admin")) {
-
-            throw new ValidationError("The username is already taken.");
-        }
-    }
-}
